FastVPINNs: Moved from src to Scirex
- Moved the core module from src to scirex
- Changes only done for FE Modules
