+++ scirex/core/sciml/FE/basis_2d_QN_Jacobi.py
@@ -1,5 +1,4 @@
-
-# Copyright (c) 2024 Zenteiq Aitech Innovations Private Limited and 
+# Copyright (c) 2024 Zenteiq Aitech Innovations Private Limited and
 # AiREX Lab, Indian Institute of Science, Bangalore.
 # All rights reserved.
 #
