+++ scirex/core/sciml/FE/basis_function_2d.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2024 Zenteiq Aitech Innovations Private Limited and 
+# Copyright (c) 2024 Zenteiq Aitech Innovations Private Limited and
 # AiREX Lab, Indian Institute of Science, Bangalore.
 # All rights reserved.
 #
@@ -20,38 +20,35 @@
 # For any clarifications or special considerations,
 # please contact: contact@scirex.org
 """
-file: basis_function_3d.py
-description: Abstract class for all basis functions 3D
-authors: Thivin Anandh D
-changelog: 03/May/2024 - First version
-known_issues: None
-dependencies: None specified.
+The file `basis_function_2d.py` contains a wrapper class for all the finite element basis functions 
+used in the FE2D code. The 2D basis functions have methods to return the value 
+of the basis function and its derivatives at the reference point (xi, eta).
+
+Author: Thivin Anandh D
+
+Changelog: 30/Aug/2023 - First version
+
+Known issues: None
+
+Dependencies: None specified
 """
 
 from abc import abstractmethod
 
 
-class BasisFunction3D:  # pragma: no cover
+class BasisFunction2D:
     """
-    Represents a basis function in 3D.
+    This class defines the basis functions for a 2D element.
 
-    Args:
-        num_shape_functions (int): The number of shape functions.
-
-    Methods:
-        value(xi, eta, zeta): Evaluates the basis function at the given xi and eta coordinates.
-        gradx(xi, eta, zeta): Computes the partial derivative of the basis function with respect to xi.
-        grady(xi, eta, zeta): Computes the partial derivative of the basis function with respect to eta.
-        gradxx(xi, eta, zeta): Computes the second partial derivative of the basis function with respect to xi.
-        gradxy(xi, eta, zeta): Computes the mixed partial derivative of the basis function with respect to xi and eta.
-        gradyy(xi, eta, zeta): Computes the second partial derivative of the basis function with respect to eta.
+    :num_shape_functions (int): The number of shape functions.
+    :value(xi, eta): Evaluates the basis function at the given xi and eta coordinates.
     """
 
     def __init__(self, num_shape_functions):
         self.num_shape_functions = num_shape_functions
 
     @abstractmethod
-    def value(self, xi, eta, zeta):
+    def value(self, xi, eta):
         """
         Evaluates the basis function at the given xi and eta coordinates.
 
@@ -60,9 +57,10 @@ class BasisFunction3D:  # pragma: no cover
         :return: The value of the basis function at the given coordinates.
         :rtype: float
         """
+        pass
 
     @abstractmethod
-    def gradx(self, xi, eta, zeta):
+    def gradx(self, xi, eta):
         """
         Computes the partial derivative of the basis function with respect to xi.
 
@@ -71,9 +69,10 @@ class BasisFunction3D:  # pragma: no cover
         :return: The partial derivative of the basis function with respect to xi.
         :rtype: float
         """
+        pass
 
     @abstractmethod
-    def grady(self, xi, eta, zeta):
+    def grady(self, xi, eta):
         """
         Computes the partial derivative of the basis function with respect to eta.
 
@@ -82,9 +81,10 @@ class BasisFunction3D:  # pragma: no cover
         :return: The partial derivative of the basis function with respect to eta.
         :rtype: float
         """
+        pass
 
     @abstractmethod
-    def gradxx(self, xi, eta, zeta):
+    def gradxx(self, xi, eta):
         """
         Computes the second partial derivative of the basis function with respect to xi.
 
@@ -93,9 +93,10 @@ class BasisFunction3D:  # pragma: no cover
         :return: The second partial derivative of the basis function with respect to xi.
         :rtype: float
         """
+        pass
 
     @abstractmethod
-    def gradxy(self, xi, eta, zeta):
+    def gradxy(self, xi, eta):
         """
         Computes the mixed partial derivative of the basis function with respect to xi and eta.
 
@@ -104,9 +105,10 @@ class BasisFunction3D:  # pragma: no cover
         :return: The mixed partial derivative of the basis function with respect to xi and eta.
         :rtype: float
         """
+        pass
 
     @abstractmethod
-    def gradyy(self, xi, eta, zeta):
+    def gradyy(self, xi, eta):
         """
         Computes the second partial derivative of the basis function with respect to eta.
 
@@ -115,3 +117,15 @@ class BasisFunction3D:  # pragma: no cover
         :return: The second partial derivative of the basis function with respect to eta.
         :rtype: float
         """
+        pass
+
+
+# ---------------- Legendre -------------------------- #
+from .basis_2d_QN_Legendre import *  # Normal Legendre from Jacobi -> J(n) = J(n-1) - J(n+1)
+from .basis_2d_QN_Legendre_Special import *  # L(n) = L(n-1) - L(n+1)
+
+# ---------------- Jacobi -------------------------- #
+from .basis_2d_QN_Jacobi import *  # Normal Jacobi
+
+# ---------------- Chebyshev -------------------------- #
+from .basis_2d_QN_Chebyshev_2 import *  # Normal Chebyshev
